conversao de mensagem atualizada para futura adaptacao
